pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.71.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "crewai>=1.1.0",
    "crewai-tools>=1.1.0",
    "fastapi>=0.119.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "loguru>=0.7.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "redis>=5.0.0",
    "requests>=2.32.5",
    "resend>=2.19.0",
    "tenacity>=8.0.0",
    "uvicorn>=0.38.0",
    "youtube-transcript-api>=1.2.3",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]
